% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonerrr-package.R
\docType{package}
\name{lonerrr-package}
\alias{lonerrr}
\alias{lonerrr-package}
\title{lonerrr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/f-dallolio/lonerrr}
  \item Report bugs at \url{https://github.com/f-dallolio/lonerrr/issues}
}

}
\author{
\strong{Maintainer}: Filippo Dall'Olio \email{fdallolio@brock.ca}

}
\keyword{internal}
