% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr-standalone.R
\name{standalone-purrr-accumulate}
\alias{standalone-purrr-accumulate}
\alias{accumulate}
\alias{accumulate_right}
\alias{detect}
\alias{detect_index}
\title{Accumulate intermediate results of a vector reduction.}
\usage{
accumulate(.x, .f, ..., .init)

accumulate_right(.x, .f, ..., .init)

detect(.x, .f, ..., .right = FALSE, .p = is_true)

detect_index(.x, .f, ..., .right = FALSE, .p = is_true)
}
\description{
Accumulate intermediate results of a vector reduction.

Find the value or position of the first match.
}
