% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr-standalone.R
\name{standalone-purrr-imap}
\alias{standalone-purrr-imap}
\alias{imap}
\alias{imap_vec}
\title{Apply a function to each element of a vector, and its index.}
\usage{
imap(.x, .f, ...)

imap_vec(.x, .f, ...)
}
\description{
Apply a function to each element of a vector, and its index.
}
