% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr-standalone.R
\name{standalone-purrr-keep}
\alias{standalone-purrr-keep}
\alias{compact}
\alias{keep}
\alias{discard}
\alias{map_if}
\title{Keep/discard elements based on their values.}
\usage{
compact(.x)

keep(.x, .f, ...)

discard(.x, .p, ...)

map_if(.x, .p, .f, ..., .else)
}
\description{
Keep/discard elements based on their values.
}
