% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr-standalone.R
\name{standalone-purrr-map2}
\alias{standalone-purrr-map2}
\alias{map2}
\alias{map2_lgl}
\alias{map2_int}
\alias{map2_dbl}
\alias{map2_chr}
\alias{map2_vec}
\title{Map over two inputs}
\usage{
map2(.x, .y, .f, ...)

map2_lgl(.x, .y, .f, ...)

map2_int(.x, .y, .f, ...)

map2_dbl(.x, .y, .f, ...)

map2_chr(.x, .y, .f, ...)

map2_vec(.x, .y, .f, ...)
}
\description{
Map over two inputs
}
