% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr-standalone.R
\name{standalone-purrr-modify}
\alias{standalone-purrr-modify}
\alias{modify}
\alias{modify_if}
\title{Modify elements selectively}
\usage{
modify(.x, .f, ...)

modify_if(.x, .p, .f, ..., .else)
}
\description{
Modify elements selectively
}
