% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr-standalone.R
\name{standalone-purrr-pmap}
\alias{standalone-purrr-pmap}
\alias{pmap}
\alias{pmap_vec}
\title{Map over multiple input simultaneously (in "parallel").}
\usage{
pmap(.l, .f, ...)

pmap_vec(.l, .f, ...)
}
\description{
Map over multiple input simultaneously (in "parallel").
}
