% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr-standalone.R
\name{standalone-purrr-reduce}
\alias{standalone-purrr-reduce}
\alias{reduce}
\alias{reduce_right}
\title{Reduce a list to a single value by iteratively applying a binary function.}
\usage{
reduce(.x, .f, ..., .init)

reduce_right(.x, .f, ..., .init)
}
\description{
Reduce a list to a single value by iteratively applying a binary function.
}
