% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr-standalone.R
\name{standalone-purrr-simplify}
\alias{standalone-purrr-simplify}
\alias{list_simplify}
\title{Simplify a list to an atomic or S3 vector.}
\usage{
list_simplify(x)
}
\description{
Simplify a list to an atomic or S3 vector.
}
